Update NearbyMeasurementCard to use 'distance' instead of 'distanceKm'
